Changed all f32's to f64's.

+++ src/geolocation/response/mod.rs
@@ -8,6 +8,6 @@ use serde::{Serialize, Deserialize};
 #[derive(Clone, Debug, PartialEq, PartialOrd, Serialize, Deserialize)]
 pub struct Response {
     pub location: Option<LatLng>,
-    pub accuracy: Option<f32>,
+    pub accuracy: Option<f64>,
     pub error: Option<ErrorObject>,
 } // struct
